Homework 2.10 fix
