--- src/main/java/qqqbbb/Homework2dot10/BadRequestException.java
@@ -1,10 +1,0 @@
-package qqqbbb.Homework2dot10;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.BAD_REQUEST, reason = "Bad request")
-public class BadRequestException extends RuntimeException
-{
-
-}
