--- src/main/java/qqqbbb/Homework2dot10/Main.java
@@ -1,20 +1,0 @@
-package qqqbbb.Homework2dot10;
-
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Main
-{
-    public static void main(String[] args)
-    {
-        SpringApplication.run(qqqbbb.Homework2dot10.Main.class, args);
-//        System.out.println( StringUtils.isAlpha(null));
-//        System.out.println(StringUtils.isAlpha(""));
-//        System.out.println(StringUtils.isAlpha("_"));
-//        System.out.println(StringUtils.isAlpha("qwe"));
-//        System.out.println(StringUtils.isAlpha("йцу"));
-    }
-
-}
