+++ src/main/java/qqqbbb/homework2dot5/EmployeeAlreadyAddedException.java
@@ -1,8 +1,9 @@
 package qqqbbb.homework2dot5;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus
+@ResponseStatus(value = HttpStatus.EXPECTATION_FAILED, reason = "Employee Already Added")
 public class EmployeeAlreadyAddedException extends RuntimeException
 {
 }
