+++ src/main/java/qqqbbb/homework2dot5/EmployeeNotFoundException.java
@@ -1,8 +1,9 @@
 package qqqbbb.homework2dot5;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus
-public class EmployeeNotFoundException extends RuntimeException
+@ResponseStatus(value = HttpStatus.EXPECTATION_FAILED, reason = "Employee Not Found")
+class EmployeeNotFoundException extends RuntimeException
 {
 }
