+++ src/main/java/qqqbbb/homework2dot5/EmployeeStorageIsFullException.java
@@ -3,6 +3,6 @@ package qqqbbb.homework2dot5;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus
-public class EmployeeStorageIsFullException extends RuntimeException
+class EmployeeStorageIsFullException extends RuntimeException
 {
 }
